[wait_until_unavailable] Stop sending analytics for oci_config::wait_until_unavaliable to bolt/puppet
The `Bolt#Executor` method `report_function_call` is meant to report *only* the functions built in to boltlib (the plan functions that ship with Bolt). Functions defined in external modules should not report back. Instead of attempting to "allowlist" a set of boltlib functions (as this is the first time this has happened as it appears to have been copy-pasted over from boltlib) this commit simply removes that method invocation from the external module.