[core] Fix name caching when single entry is added to empty cache
When a single antry was added to the cache before the cache was filled, the cache was still  marked as filled. This caused name lookups to fail. This commit only marks the type as cached whenit is really filled. Single entries added to the cache leave the time markings untouched.